chore: ui, 기능 분리
